Fixing my new test for high #cores machines
My new test failed last night on machines with high node counts
because if #numtasks * chunksize > problemsize the dynamic iterator
is serialized.  I fixed this by dialing down dataParTasksPerLocale
since that will determine the default number of tasks used (which
is what I was trying to exercise) and keep it from hitting this
serialization condition.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20996 3a8e244f-b0f2-452b-bcba-4c88e055c3ca