Unbreak start_test.
[Reviewed by Vass.]

When I "fixed" start_test with r20988 I missed the fact that one of the
platform references, namely the location of the compiler, really did need to
be the host architecture.  By changing it to be the target architecture, I
broke the script when the two were different and a compiler for the target
architecture had not coincidentally already been built.  This change fixes
that problem.  The script also now reports both the host and target platforms,
since both are important.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20999 3a8e244f-b0f2-452b-bcba-4c88e055c3ca