Make sizeof() not supported by default, once again
In r20995, the compound statement in the generated SysCTypes.chpl that
protected the extern definition of sizeof() from leaking out to all Chapel
code was removed, the rationale being that the file-level compound statement
seemed to be getting interpreted as a module.  Having reviewed this, I don't
believe this is the case -- it seems to be treated as a compound statement
within the module's init function as you'd expect.  My guess is that the
misdiagnosis may have come from test/extern/sungeun/sizeof.chpl which
introduces a module named 'sizeof' based on the filename and tha this may
have been interpreted as having been created by the compound statement in
SysCTypes.chpl (at least that's the only explanation I can come up with).

Historically, we've avoided introducing sizeof() into Chapel due to
confusion about what it should do (our confusion / potential user
confusion).  In particular, when arrays are defined using, say, a
stack-allocated descriptor pointing to a dynamically allocated memory
segment (or perhaps multiple segments), it's not clear what sizeof() should
refer to.  Rather than choosing a definition that may not be what a user
wants/expects, we've (so far) chosen to avoid defining it.  Part of the
rationale also relies on the argument that most users shouldn't need to know
how large Chapel types are, and have other means of doing so when necessary
(numBits, .size routines, etc).

This commit reintroduces the compound statement, adds an additional extern
sizeof in DefaultRectangular to another recent case where it started being
used, and adds a test to lock in this behavior going forward to avoid having
it slip back in unnoticed (a version of the rationale in the previous
paragraph appears in that test as well).  extern sizeofs() are decorated
with pragma "no prototype" in order to avoid conflicts when using the
--gen-extern-prototypes flag since sizeof() is an operator in C.

The motivation for this stems from a user query expressing confusion about
what they were getting from sizeof() which is the first I'd noticed that it
had slipped back into scope by default.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21015 3a8e244f-b0f2-452b-bcba-4c88e055c3ca