Make LULESH support 1D vs. 3D data structures via a config param
This commit changes our LULESH implementation to support the
general/irregular 1D data structures that we have been in
all recent versions; or to use 3D domains/arrays when running
using the computed cube input set.  This was a great example
of the separation of data structure implementation from code
in Chapel -- the diff to lulesh.chpl is small and is predominantly
due to formatting changes (whitespace changes, variable renaming).
The intellectual parts were as follows:

1) Add a new config param:

              use3DRepresentation = false,


2) Redefine the two main domains from 1D:

const ElemSpace = {0..#numElems},
      NodeSpace = {0..#numNodes};

to "3D or 1D" based on the config param:

const ElemSpace = if use3DRepresentation
                    then {0..#elemsPerEdge, 0..#elemsPerEdge, 0..#elemsPerEdge}
                    else {0..#numElems},
      NodeSpace = if use3DRepresentation
                    then {0..#nodesPerEdge, 0..#nodesPerEdge, 0..#nodesPerEdge}
                    else {0..#numNodes};

3) Insert an error in the event of trying to run in a nonsensical mode:

if (use3DRepresentation && (luleshInit.filename != "")) then
  halt("The 3D representation does not support reading input from files");

Period.  All the physics remained identical. Not bad for a few hours
work!

Bigger changes were required in the luleshInit module because it
computes the cube dataset based on the indices, so had to be
generalized to know how to work with either 1D or 3D indices.  This
was fairly straightforward -- the main challenge was determining how
best to determine whether we were running the 3D or 1D code.  I wanted
to do this without referring to the config param so that the
initialization module wouldn't have a dependence back on lulesh.chpl
(and I didn't want to move the config param into luleshInit.chpl, just
to keep them all together).

It's worth noting that I think this version continues to be a little
different than the original Sung version in that I believe that one
would compute neighboring elements on the fly by taking an (i,j,k)
coordinate and adding/subtracting one to it.  This version explicitly
stores the list of neighbors, as in the irregular case, which I
believe is also what the reference version does.  So I would argue for
sticking with the current approach (rather than special casing the
code for finding neighbors in 3D) and retiring Sung's studies version
of LULESH as a result (we were keeping it around primarily to keep a
working-ish copy of the 3D version around).


Other changes:

* I renamed sparseMaterials to useSparseMaterials to make it more
  similar to the other config params.  This required also changing
  the .compopts file.

* I needed a separate set of .execopts and .goods for testing the 3D
  cases compared to the 1D cases, so created a second test --
  test3DLulesh.chpl -- that 'use's the original test and runs it in
  the different mode to do this.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21020 3a8e244f-b0f2-452b-bcba-4c88e055c3ca