Eliminate mysterious re-linking of the chpl executable. [trivial]
In the rule
    $(CHPL) : $(CHPL_BIN_DIR) $(CHPL_OBJS)
the date on $(CHPL_BIN_DIR) is updated every time $(CHPL) is rebuilt and it is
always slightly newer.  As a result, $(CHPL) is remade each time.

The problem is solved in GNU make by supplying $(CHPL_BIN_DIR) as an
"order-only" dependency, i.e. one that is listed to the right of a vertical bar.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21023 3a8e244f-b0f2-452b-bcba-4c88e055c3ca