Fixing a ssize_t/int(64) mismatch on 32-bit platforms
Akin to other QIO issues we've seen, an int(64) was being passed
to a ssize_t argument causing a mismatch on platforms where
ssize_t is an int(32).  Something about my new test3DLulesh.chpl
test triggered this case.  I inserted a cast to deal with it,
as we've done with others.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21025 3a8e244f-b0f2-452b-bcba-4c88e055c3ca