Allow main to handle arguments or return an integer.
[reviewed by Tom, feature reviewed by Brad]

This patch adds the ability for a user to opt-in for explicit argument
handling if they declare main to take in an array of strings. Since many
existing config vars are important or essential (e.g. -nl or --numLocales),
only arguments that were not already handled are passed as arguments to the
declared main.  Also, if -- is specified on the command line, all arguments
after it go to the user main. In this way, it would be possible to replace
some legacy program taking arbitrary arguments with

 ./my_prog -- $OLDARGS


Similarly, if main is declared to return an int, this patch makes the Chapel
runtime complete normal program termination and return the value. When main is
declared with no arguments and no return type, the previous behavior is
preserved.  For more details about the capability this patch provides, see the
handy new READMEs.


I've had some discussion with Brad about a module-oriented capability for
handling arbitrary arguments. We agree that might be desireable (although I
think this patch might be enough). The concept would be that module-argument
handling would look something like this in the main argument processing loop:

 for(curArg=1; curArg<argc; curArg+=handled) {
   for each module m {
      handled = ask module m to consume arguments starting at argv[curArg],
                it returns the number of arguments consumed, or 0
                if it opts not to handle any of them.
      if( handled > 0 ) break;
   }
 }

See our discussion on chapel-developers entitled "review - argument handling".

Future work:

* add support for config variables of aggregate types (e.g. tuples, arrays)

* provide a module-oriented capability for creating a --help message via a
   well-defined entry point (e.g., proc help();}

* (pending module-oriented argument handling) have modules register their
   intent to handle the --help flag

* consider providing compiler flags to select between "higher
   productivity" vs. "more manual" handling of --help/-h

* consider allowing a Chapel programer to change the meaning of certain special
  flags (e.g. -h -q -v -b -t --gdb) in their program.

 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21030 3a8e244f-b0f2-452b-bcba-4c88e055c3ca