Reimplement locale-aware tasking in a slightly less kludgy manner.
The main improvements are:
 - No longer assuming that there is an array of nodes at the top level;
 - Factoring out the RootLocale class; and
 - Providing backward-compatible .id and .name functions which can still have
   architecture-dependent behavior.

Details
-------

    test/multilocale/bradc/locales/queryLocaleLocation.chpl
    test/multilocale/bradc/locales/queryLocaleLocation.good
    test/multilocale/bradc/locales/queryLocaleLocation.comm-none.good
User code should not know about chpl_id, so that portion of the test is removed.

    modules/internal/DefaultArchitecture.chpl
DefaultRootLocale is derived from RootLocale, so it can support the
getDefaultLocaleSpace(), getDefaultLocaleArray() and localeIDtoLocale()
interfaces.
chpl_id() and chpl_name() are implemented, to provide
architecture-dependent definitions for the .id and .name members of class
locale.
_node_id is the new id in readWriteThis().
Provided taskInit() and taskExit() default implementations.
 
    modules/internal/RootLocale.chpl
DefaultRootLocale now derives from RootLocale, so this module cannot depend on
DefaultArchitecture.
Defined new dynamically-bound functions getDefaultLocaleSpace(),
getDefaultLocaleArray() for backward-compatible support of Locales[] and
LocaleSpace.
Generalized chpl_localeID_to_locale() so it delegates to the RootLocale
implementation rather than just selecting a child at that level.
Moved the definitions for Locales[] and LocaleSpace into
DefaultArchitecture.chpl, to force the use of rootLocale.getDefaultLocaleArray()
and rootLocale.getDefaultLocaleSpace() going forward.

    modules/internal/ChapelLocale.chpl
Cannot depend on DefaultRectangular.
Moved emptyLocaleSpace &c to RootLocale.chpl.
Allow locales to be created (for the nonce).
Add a parent field.
Define statically-linked .id and .name in terms of dynamically-linked chpl_id()
and chpl_name(), resp.
Removed the getChildSpace() interface, to see if we can do without it.
Moved getChildArray() into RootLocale (and renamed as getDefaultLocaleArray()).
Added interfaces for taskInit() and taskExit().  These could be pure virtual,
except that we are using them in early calls to the default locale instance.
rootLocale is defined as nil and _here is initialized to an instance of locale()
to satisfy early usage.
chpl_getPrivatizedCopy moved up to remain close to the definition of _here.

    modules/internal/ChapelBase.chpl
Defined the convenience function _throwPVFCError() to signal a pure virtual
function call error.  Factored this out of ChapelLocale.chpl and RootLocale.chpl
(potentially saves string space in the compiled image).

    compiler/include/primitive.h
    compiler/AST/primitive.cpp
    compiler/AST/expr.cpp
    compiler/optimizations/optimizeOnClauses.cpp
Added PRIM_LOCALE_ID.

    compiler/AST/build.cpp
Removed the x.locale.id optimization.
Renamed chpl_int_to_locale as chpl_localeID_to_locale.

    compiler/passes/cleanup.cpp
Added addTaskMgtInOnBlocks() to call taskInit() and taskExit() w.r.t. the
current locale at the beginning and end of each on block, respectively.



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/hierarchical_locales@21032 3a8e244f-b0f2-452b-bcba-4c88e055c3ca