Removed these unused functions from our utility classes in the compiler:
   Map::some_disjunction(Map<K,C> &m)
   Vec::set_remove(C a)
   Vec::set_intersection(Vec<C,S> &vv)
   Vec::some_intersection(Vec<C,S> &vv)
   Vec::some_disjunction(Vec<C,S> &vv)
   Vec::set_intersection(Vec<C,S> &vv, Vec<C,S> &result)
   Vec::set_disjunction(Vec<C,S> &vv, Vec<C,S> &result)
   Vec::set_difference(Vec<C,S> &vv, Vec<C,S> &result)
   Vec::some_difference(Vec<C,S> &vv)
   Vec::set_count()

[reviewed by Tom]

Removed only the declarations/definitions - there were no calls to these.
Almost removed Vec::move, but there is one use of it in codegen.cpp.
We *can* remove Vec::remove(int index) - it is unused, too.

Motivation: my experimental use of set_remove() did not behave as expected.
(See below for details.)  While, on the surface, its implementation
looks simple and flawless, I chose not to investigate.
Instead I observed that set_remove() is not used anywhere in the compiler.
So, should there be a problem with it, we would not have noticed.

In the interest of time, and since we would like to switch to std::set
for our set needs, there was the group consensus that we can remove
set_remove() and other unused methods from the code base.
This prevents others from tripping over any problems
that they [the methods, that is] may have.

Here is an overview of my experiment, just in case. In this function:

   flattenNestedFunctions(Vec<FnSymbol*>& nestedFunctions)

there is a loop forv_Vec(FnSymbol, fn, nestedFunctions)
inside which there are additions to nestedFunctionSet.
At the end of that loop I added this code:

   if ((*fn->calledBy).n == 0) {
     bool before = nestedFunctionSet.set_in(vasstest);
     nestedFunctionSet.set_remove(fn);
     bool after  = nestedFunctionSet.set_in(vasstest);
   }

where 'vasstest' is some FnSymbol* that was in nestedFunctionSet
the first time my additional code was executed.

By printing out 'before' and 'after', I observed that
(a) when fn==vasstest, set_remove(fn) does not remove vasstest;
(b) when fn!=vasstest, set_remove(fn) removes vasstest.
This is the opposite of what should happen.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21040 3a8e244f-b0f2-452b-bcba-4c88e055c3ca