Store here ID in task-private data. Make the runtime ignorant of the layout of chpl_localeID_t.  (The runtime treats it more-or-less opaquely as a 64-bit integer.)
Details:
    runtime/include/chpl-tasks.h
    runtime/src/tasks/none/tasks-none.c
    runtime/src/tasks/mta/tasks-mta.c
    runtime/src/tasks/fifo/tasks-fifo.c
    compiler/include/primitive.h
    compiler/AST/primitive.cpp
    compiler/AST/expr.cpp
    compiler/passes/parallel.cpp
    compiler/optimizations/optimizeOnClauses.cpp
Converted set- and get-subloc routines to store and retrieve the whole
locale ID instead.

    runtime/include/chplcgfns.h
    runtime/include/chplcgtypes.h
    runtime/src/tasks/massivethreads/tasks-massivethreads.c
Move runtime-only definition of chpl_localeID_t into chplcgtypes.h.

    runtime/include/stdchpl.h
    runtime/include/chpl-gen-includes.h
    compiler/passes/codegen.cpp
Move chpl-comm-compiler-macros down so it can use the
compiler-generated definition of chpl_localeID_t.

    runtime/include/chpl-comm-compiler-macros.h
Add chpl_gen_getLocaleID() and chpl_get_setLocaleID() to convert
between the compilers and runtimes understanding of what a locale ID is.

    runtime/include/chpltypes.h
c_locale_t is now a 64-bit integer, as far as the runtime is concerned.

    runtime/src/chpltypes.c
Use chpl_gen_getLocaleID() where the current locale ID is needed.
(This is in contrast to assembling it from the node ID and the subloc ID.

    modules/internal/DefaultArchitecture.chpl
Use the node name with the gasnet node ID appended if GASNET_SPAWNFN=="L".
 


git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/hierarchical_locales@21042 3a8e244f-b0f2-452b-bcba-4c88e055c3ca