A more general implementation of the bulkTransferStride optimization.
This new approach Vass devised during his research stay in Malaga. Essentially,
it makes the implementation of the message aggregation more independent of the
actual distribution so that a new data distribution module can be enhanced with
bulk communications without knowing the details of the other already
available/implemented data distributions.

Alberto implemented it, so it's now dealing with Block and Cyclic distributions
and it is working fine for several tests doing slice assignments for
block-to-block, cyclic-to-cyclic, block-to-cyclic and cyclic-to-block cases. We
are now conducting some experiments on raven Cray machine.

For the record, this is the mail that Vass sent us with the sketch of his idea:

********************************************************************************
Domain Map Interface for Bulk Communication

(A) Each domain map provides the following operations:

  - void GlobalArr.doiBulkTransferTo(DR)
  - void GlobalArr.doiBulkTransferFrom(DR)

where:
   - GlobalArr is the global array descriptor for that domain map,
   - DR is a default rectangular array. DR can be on any locale.

Discussion

What we really want is to be able to do the assignments A=DR and DR=A, where A
is an array under the domain map in question and DR is a default rectangular
array. But a domain map cannot (should not?) implement = directly. So instead we
require the above two methods. We will need to modify the modules so that A=DR
and DR=A redirect to the above methods.

Why two methods (vs. just one in the current implementation)? It is the
underlying idea of this interface that each domain map ensures that A=DR and
DR=A are implemented. (This idea is open for discussion of course.) Therefore
the implementations belong in the corresponding domain maps. Correspondingly, we
put the methods in the GlobalArr class and give them different names to
distinguish the direction.

(B) Each domain map provides the following operations:

   - void GlobalArr.doiBulkTransferTo(B)
   - void GlobalArr.doiBulkTransferFrom(B)

where
   B is an array of any domain map.

Discussion:

The idea is that each GlobalArr will commonly be a collection of DR arrays.
(Once we have other cases, we will reconsider the interface.) Therefore
GlobalArr.doiBulkTransferTo(B) can be implemented by invoking a set of
(B[SLC_i])._value.doiBulkTransferFrom(DR_i) where SLC_i and DR_i are computed
for each element of the set. Analogously for the other direction.

For that, GlobalArr.doiBulkTransferTo(B) will:
* go over its DR_i arrays (e.g. via a coforall); 
and for each DR
* compute SLC_i, which is the part of B to which the values of DR must be copied.

************************************************************************************

*** An example: Block <-- Cyclic (on 2 locales)

n=8;
const S = {1..n}; 
const CDom: domain(1) dmapped Cyclic(startIdx=S.low)=S;
var Dist1 = new dmap(new Block({1..n}));
var Dom1: domain(1,int) dmapped Dist1 = {1..n};
var B:[Dom1] int(64);
var C:[CDom] int(64);

B[1..7 by 2] = C[1…7 by 2]  ==> [ L0,L0,L1,L1] = [L0,L0,L0,L0]

B.doiBulkTransfer(C)
	B[1..7 by 2].doiBulkTransferFrom(C[1..7 by 2])
    	       C[1..3 by 2].doiBulkTransferToDR(B[1..3 by 2]);
    	       	      	 B[1..3 by 2].doiBulkTransfer(C[1..3 by 2]); // On L0: DR<-DR
 
		C[5..7 by 2].doiBulkTransferToDR(C[5..7 by 2]);
    		       	  B[5..7 by 2].doiBulkTransfer(C[5..7 by 2]); // On L1: DR<-DR 

**** More details ****

test/optimizations/bulkcomm/alberto/:
	Two new folders: Block and Cyclic, for the new corresponding tests.
Several tests for all rhs=lhs cases, due to rhs and lhs can be
Block, Cyclic or Default Rectangular 

modules/internal/ChapelArray.chpl:
	Updated so as to follow the doiBulkTransferStride path for Cyclic,
Block and DR array assignments.

modules/internal/ChapelDistribution.chpl:
	New method to check whether or not an array is Cyclicly
	distributed.

modules/internal/DefaultRectangular.chpl:
	New method
	DefaultRectangularArr.dsiSupportsBulkTransferStride() to
	verify that the distribution supports "TransferStride"
	Corrected minor bug when calling doiBulkTransferStrideComm().

modules/dists/BlockDist.chpl
	New methods:
	+proc BlockArr.doiBulkTransferTo(Barg) --> where Barg is Block or Cyclic 
	+proc BlockArr.doiBulkTransferToDR(Barg) --> where Barg is DR
	+proc BlockArr.doiBulkTransferFrom(Barg) --> where Barg is Block or Cyclic 
	+proc BlockArr.doiBulkTransferFromDR(Barg) --> where Barg is DR
	+proc BlockArr.dsiSupportsBulkTransferStride() to verify that the
	distribution supports "TransferStride".

modules/dists/CyclicDist.chpl
	Añadido los métodos:
	+proc CyclicArr.doiBulkTransferTo(Barg) --> where Barg is Block or Cyclic 
	+proc CyclicArr.doiBulkTransferToDR(Barg) --> where Barg is DR
	+proc CyclicArr.doiBulkTransferFrom(Barg) --> where Barg is Block or Cyclic 
	+proc CyclicArr.doiBulkTransferFromDR(Barg) --> where Barg is DR
	+proc CyclicArr.dsiSupportsBulkTransferStride() to verify that the
	distribution supports "TransferStride". 



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/bulkComms@21043 3a8e244f-b0f2-452b-bcba-4c88e055c3ca