Miscellaneous minor improvements/cleanups in compiler, modules.
[reviewed by David]

* Made the former "accessing the locale of a local expression" warning
be more correct w.r.t. the nature of the issue.
Adjusted .good correspondingly.

* Made the line number corresponding to a promotion wrapper be that
of the callsite, not the callee. This was approved by the group over email.

* Made cleanFilename(BaseAST*) more robust w.r.t. null pointers.
This was needed for its use in a compiler error message that I encountered.
Also, use the node's filename whenever it's available, even if
'yyfilename' is also available.

* Removed a duplicate 'use MemTracking' statement.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21049 3a8e244f-b0f2-452b-bcba-4c88e055c3ca