Updating .skipif for this future test -- we've been getting lucky
This test started passing on Feb 8th, but it seems we've just been
getting lucky.  I can only get it to fail using valgrind currently,
so I limited the skipif to that case for now.

Unfortunately, I don't think we test futures with valgrind, so we'll
have to review this one manually, perhaps after strings get
rewritten?



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21050 3a8e244f-b0f2-452b-bcba-4c88e055c3ca