Adding Qthreads version 1.9 to our repository
This commit adds the new Qthreads version 1.9 to our repository
and makes it the default version of Qthreads used (previously,
we were using 1.8).  Spot testing makes it appear that this is
working at least as well as 1.8 (Dylan Stark has also done
testing indicating this), but I'll let it run for a night before
removing version 1.8.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21051 3a8e244f-b0f2-452b-bcba-4c88e055c3ca