Retiring this future
This future was designed to show that we needed a task pool that
didn't block on sync variable operations to handle arbitrarily
many threads.  We now have such task layers, and this test's
failure/passage depends a lot on how such things get scheduled,
so it seems that it isn't needed as a placeholder any longer.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21054 3a8e244f-b0f2-452b-bcba-4c88e055c3ca