Removing --fast flags from lulesh correctness tests
As stated in r20802, my thinking is that we should
remove --fast from correctness tests in order to find
incorrectnesses and not time out in compilation.
Lulesh has been compiled with --fast for quite awhile
now, but the 3D version has recently been timing out
in compilation.  Removing the --fast (and the 10 minute
timeout file) in hopes of resolving this.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21063 3a8e244f-b0f2-452b-bcba-4c88e055c3ca