Add locale lookup optimization.
When the object passed in an on statement is a locale object, it is unnecessary
to look it up via its locale ID, so this extra dereference operation is removed.

Details:

    compiler/include/misc.h
Added INT_ASSERT_AND_RETURN_NULL macro.

    compiler/include/type.h
    compiler/AST/type.cpp
Added isSubClass() predicate.

    compiler/include/vec.h
Added only() accessor.

    compiler/include/astutil.h
    compiler/AST/astutil.cpp
Added buildDefUseMaps(BlockStmt*, etc.).
Made buildDefUseMaps(Vec<Symbol*>&, etc) inline.

    compiler/passes/parallel.cpp
Added the locale lookup optimization here.
In each on block, we look for an on expression matching the pattern
  onExpr = chpl_localeID_to_locale(__primitive("_wide_get_locale", x)),
where x is of a type derived from dtLocale.
When there is a match, this is replaced by
  onExpr = x
.


git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/hierarchical_locales@21079 3a8e244f-b0f2-452b-bcba-4c88e055c3ca