Trying to reduce the chances of failure for this test
This test has continued to fail recently due to verification issues.
In speaking with Greg this morning, I was lamenting that in spite
of dialing the problem size and number of updates to match our other
recent failures, we were still seeing them here, and he reminded me
about the clustering of the random numbers which caused me to remember
that for the release version we switched to the other random number
generator for testing purposes, but probably had not done so here.
Now we have.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21081 3a8e244f-b0f2-452b-bcba-4c88e055c3ca