Move locale ID setup inside the wrapon_fn.
Preparatory to moving the locale lookup inside the local test, the code for
storing and restoring a temporary copy of the "here" locale and setting up the
new one when entering an on block is moved from the "on" function itself into
the wrapon_fn.  This means that these steps are not done if the local test
succeeds and the "on" function is run locally.  

Caveat: In this case "local" means the exact same locale ID including the
sublocale ID.



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/hierarchical_locales@21092 3a8e244f-b0f2-452b-bcba-4c88e055c3ca