Use explicit processor atomics for reference counting distributions, domains, and arrays
[reviewed by gbt]

This commit changes the atomics used for reference counting to be
explicit processor atomics.

It seems that in the common case, accesses to the reference count are
from the same locale as the reference counted object was allocated on.
Therefore, use processor atomics even when network atomics are
available.

This caused the test atomicflag.chpl to run into the "module name
conflicts with some other name" problem, so I changed the test name to
atomic_bool.chpl, which is a better name anyways.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21095 3a8e244f-b0f2-452b-bcba-4c88e055c3ca