Fixing a regression in the analysis of memory leaks script
At some point, it seems that our output for the memory leaks
runs changed to the extent that the perl wasn't correctly
finding the program name anymore.  This is the reason that
the list of largest offending programs has been missing in
the memory leaks logs for some time now.

This change fixes the problem by adding an extra space after
the trailing .chpl, which found its way in... sometime.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21096 3a8e244f-b0f2-452b-bcba-4c88e055c3ca