Minor housekeeping to reduce compilation warnings/errors
[Reviewed by Brad]

These fixes will hopefully reduce errors on XMT.

sys_writev and friends should all be taking in a const struct iovec*,
but not all system calls take in a const so we added a cast.

Also fixed sys_page_size() to be written in a way to avoid
unreachable statements.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21097 3a8e244f-b0f2-452b-bcba-4c88e055c3ca