Added a bug future where a compiler crashes instead of generating a user error.
This is a trimmed-down version of a test case submitted by Kellen Donohue,
who is a student in Brad's PMP 524 course.

The crash happens with the "the type of the actual argument '_unknown' is
unknown or generic" message that perhaps has been captured elsewhere.
However, this one occurs in a peculiar scenario, so I feel it's worth
capturing separately.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21098 3a8e244f-b0f2-452b-bcba-4c88e055c3ca