Remove locale argument from the called on_fn.
    runtime/include/chpl-gen-includes.h
    compiler/include/primitive.h
    compiler/AST/primitive.cpp
    compiler/AST/expr.cpp
    compiler/optimizations/optimizeOnClauses.cpp
Added PRIM_IS_HERE and chpl_is_here() to help with generating the "here" test.

    runtime/include/chpl-comm-compiler-macros.h
Added new inlines chpl_is_node_local() and rewrote chpl_test_local() to use it.
This clarifies what is meant by "local", as pertainint to node ids, and not the
localeID as a whole.

    compiler/include/baseAST.h
Changed some macros to inlines.

    compiler/passes/normalize.cpp
Rewrote some expressions using getFormal().

    compiler/passes/parallel.cpp
A wrapon_fn now steals the _dummy_locale_expr from the corresponding on_fn,
because the on_fn no longer expects the "on" locale pointer to be passed in.  It
can obtain that pointer using PRIM_GET_HERE.
Renamed _locale_arg back to _dummy_locale_arg, because only the wrapon_fn knows
about it, and even that is removed at codegen.
Changed the test for node IDs being equal to locale IDs being equal.  This is a
stronger test (but only when sublocales are used), so there should be a
redundant test in the runtime fork functions to not fork blocking on functions
that are being run on the same node.



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/hierarchical_locales@21099 3a8e244f-b0f2-452b-bcba-4c88e055c3ca