Optimization: Look inside onExpr to extract the target localeID.
This is preparatory to moving calls to chpl_localeID_to_locale() inside the
unoptimized branch of the if (prim_is_here()) test in on block calls.

This involved changing the definition of the chpl_on_locale_num
primitive, so it now takes a record of type chpl_localeID_t as its
operand.



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/hierarchical_locales@21102 3a8e244f-b0f2-452b-bcba-4c88e055c3ca