Convert is<ast_type) and to<ast_type> macros to static inline functions.
Motivation:
The is<ast_type> macros evaluate their argument twice and the to<ast_type>
macros evaluate their arguments two or three times, which can give surprising
results to someone expecting the macros to behave like functions.

Design:
The solution, of course, is to turn the macros into functions.

To save typing and reduce the chances for error, I created a template for static
inline is<ast_type> and to<ast_type> functions as a macro, and then instantiated
these two macros for all known ast_types.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21103 3a8e244f-b0f2-452b-bcba-4c88e055c3ca