Streamlined shape checking upon array assignment and minor tweaks.
[discussed over email]

All changes are in ChapelArray.chpl:

* Upon an assignment between two arrays, when boundsChecking,
verify that the two arrays are of the same shape. That is,
ensure that in each dimension the extent is the same in both arrays,
otherwise halt().

* When !boundsChecking, do not check shapes to decide whether to do
the bulk transfer optimization. The rationale is that either
the shapes match, or the assignment is not legitimate at all.

* The above is done only when both arrays are rectangular.
Otherwise, no checking is performed. Should there be any?

* Removed the default proc chpl__useBulkTransfer(a: [], b)
because it is never invoked.

* Enabled the "bulk transfer did not happen" debugging printout,
if debugBulkTransfer. It used to be disabled because the name 'writeln'
clashed with like-named module for the test program writeln.chpl.
I worked around that using 'stdout.writeln'.
(I suspect that this issue is not unique to 'writeln',
but we currently don't hit it for other names.)

* Added/adjusted a couple of comments.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21105 3a8e244f-b0f2-452b-bcba-4c88e055c3ca