Adds close calls for bits tests.
When trying to understand why this example leaks so
much memory, I realized that sometimes we're not
even calling channel.close(). It's not the source of
most of the leaks (most leaks are coming from the
channel structure not being freed even though it is
closed).

Anyway, this adds the close calls for the last example.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21112 3a8e244f-b0f2-452b-bcba-4c88e055c3ca