Corrections after conducting more tests
There was a bug in DefaultRectangular that showed up when the array
assignment involves a single element that has to be
communicated. Fixed. Anyway, we should not call the bulkComms
functions when the amount of data is less than ~100 bytes (the current
bulkComms overhead is not paid off for such small volume of data). The
latter will be taken into consideration soon.

We ran into another problem when compiling tests that use
ReplicatedDist. BulkComms mod is now implemented just for
DefaulRectangular, BlockDist and CyclicDist, so it should not be
used when assigning other kind of arrays. This is managed thanks to
dsiSupportsBulkTransferStride() that only returns true for these three
cases. However, and we don't know why, for ReplicatedDist, this
condition it's not enough and  doiBulkTransferStride
functions --that have not been defined for this distribution-- end up
being called. 

To temporally solve this issue, we have apply this ugly patch:

In ChapelArray.chpl, line 1701:

-    if chpl__isArray(b) && !chpl__serializeAssignment(a, b) {
+    if chpl__isArray(b) && !chpl__serializeAssignment(a, b) && (!a._value.isReplicatedDist() && !b._value.isReplicatedDist()){

and in ReplicatedDist.chpl, line 700:

+proc ReplicatedArr.isReplicatedDist() param {return true;}

But it should be a cleaner/clever way to achieve the same.



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/bulkComms@21118 3a8e244f-b0f2-452b-bcba-4c88e055c3ca