Add support for extern C blocks.
[Thanks to Jessica Jueckstock for her work on this patch,
 thanks to Tom for code reviewing it,
 and thanks to pretty much everybody for design discussions ]

After this patch, you can put C code in a .chpl file in an extern block
like this:

 extern {
   #include "something.h"
   static int myInt = 7;
   ....
 }

The extern { } statement "magically" adds the top-level C statements to the
enclosing module (like one might manually using existing 'extern' stuff
documented in README.extern), so that you can then do things like

 writeln(myInt);

in the same module. If you don't want to have lots of C symbols cluttering up
your namespace, it's easy to put the C in its place:

 module C { extern {
   static int foo(int x) { return x + 1; }
   ... c code here...
 } }

 writeln(C.foo(3));

This "magic" should work with C global variables, functions, structures, and
typedefs. You'll need to compile with CHPL_LLVM=llvm in order for it to work.

--- Implementation Notes ---

First, there are changes to the lexer and the parser. The lexer now has an
extra mode - called externmode. The idea is that after seeing 'extern', the
lexer is in externmode. If it then sees '{', it snarfs up everything until the
matching '}'. If it sees anything else, it leaves externmode. So from the
parser's point of view, extern { .... } turns into TEXTERN EXTERNCODE.

(Note that a side effect of some refactoring in the lexer is that
 non-ascii characters in comments will no longer be \x-escaped,
 since the \x-escaping only makes sense for string literals).
 
The parser gets a new kind of statement - extern_block_stmt - which handles
TEXTERN EXTERNCODE by making a new AST element, ExternBlockStmt.  Each
ExternBlockStmt just stores a string of the parsed code.

Then, in the readExternC pass, all ExternBlockStmts in a given module are
concatenated together and then parsed by clang. The resulting top level
declarations are put in a hashtable inside the ModuleSymbol for that module
for later use. The C code is saved into some temporary files that get compiled
with everything else during makeBinary.

(Note that all the extern C code ends up in a included at the beginning of the
C code we compile - which is currently in one giant chunk.)

Now, to enable use of symbols in the extern C blocks, we have modified
scopeResolve. In particular, if a symbol remains unresolved after scopeResolve
does its best with it, we call functions in externCResolve.cpp to translate
the C AST elements we made a note of into extern declarations, and then retry
the scopeResolve call. (this retrying is what caused the minor refactoring).
It would of course be possible to add all C symbols to the Chapel AST eagerly,
but we believe that is a bad idea because it might add lots of unused garbage
to the Chapel AST; it might make bugs in the C-to-Chapel translator prevent
any use of header files; and it might make name conflicts more common with the
C integration.

C pointers are converted into _ddata arrays; that is int* would translate into
_ddata(int). It might seems like ref(int) would be a better choice - but then
we wouldn't be able to handle multiple levels of indirection or the current
state of affairs in C where ptr[1] is just as reasonable as *ptr (barring
any other information).

--- Future Work ---

 - add a more reasonable name for _ddata (which is how C pointers are
   currently wrapped), so that error messages are not obtuse.

 - add a user-visible function to get a _ddata out of an lvalue, like addrOf
   or ptrTo. (The implementation is really easy; see
   test/extern/jjueckstock/ptr.{h,chpl} )

 - Allow length-0 tuples for variable-argument functions so that
   printf("Hello\n") can work in Chapel (and lots of other reasons).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21119 3a8e244f-b0f2-452b-bcba-4c88e055c3ca